In the pluto-container/src/main/java/org/apache/pluto/Constants.java class is a new eventNumber declared. There are also constants for max_events_size and thread_delay_time. You need this to break the process if you have loops or something like this. For a short overview look on:
http://issues.apache.org/jira/browse/PLUTO-267

api change:
XSD: 
  - replaced expiration-cache with caching
- API: changed expiration time from long to int
- removed serveFragment
- allow cookies to be set in action and event response too
- changed shared render parameters into public render parameters
- moved nextPossiblePortletModes to the render response
- added container runtime option API
- added runtime options: javax.portlet.renderHeaders, javax.portlet.includedPortletSessionScope
- renamed ClientHttpRequest into ClientDataRequest
- added MimeResponse
- clearified PortletURL.setSecure

git-svn-id: https://svn.apache.org/repos/asf/portals/pluto/branches/1.1-286-COMPATIBILITY@541884 13f79535-47bb-0310-9956-ffa450edef68

+++ portlet2-api/src/main/java/javax/portlet/CacheControl.java
@@ -24,10 +24,9 @@
  * Community Process. In order to remain compliant with the specification
  * DO NOT add / change / or delete method signatures!
  */
-/*
- * Copyright 2006 IBM Corporation.
- *
- */
+/**
+  * Copyright 2006 IBM Corporation.
+  */
 package javax.portlet;
 
 /**
@@ -52,7 +51,7 @@ public interface CacheControl {
      *          or <code>null</code> if no expiration time
      *          is set.
      */
-    public long getExpirationTime();
+    public int getExpirationTime();
     
     /**
      * Sets a new expiration time for the current response
@@ -66,7 +65,7 @@ public interface CacheControl {
      * 
      * @param time  expiration time in seconds
      */
-    public void setExpirationTime(long time);
+    public void setExpirationTime(int time);
     
     
     /**
